+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. The format 
 - Keyboard navigation for slider component [(#39)](https://github.com/btzr-io/Villain/issues/39)
 - Preview mode [(#27)](https://github.com/btzr-io/Villain/issues/27)
 - Manga mode [(#10)](https://github.com/btzr-io/Villain/issues/10)
+- Keyboard shortcuts [(#20)](https://github.com/btzr-io/Villain/issues/20)
 
 ### Fixed